Refactored the forecast to get json to better match requested
