+++ app/poros/daily_weather.rb
@@ -8,12 +8,17 @@ class DailyWeather
               :icon
 
   def initialize(argument)
-    @date = argument[:dt]
-    @sunrise = argument[:sunrise]
-    @sunset = argument[:sunset]
+    @date = date_format(argument[:dt])
+    @sunrise = date_format(argument[:sunrise])
+    @sunset = date_format(argument[:sunset])
     @max_temp = argument[:temp][:max]
     @min_temp = argument[:temp][:min]
     @conditions = argument[:weather].first[:description]
     @icon = argument[:weather].first[:icon]
   end
+
+    def date_format(utc)
+      date = Time.zone.at(utc).strftime('%b %e, %l:%M %p')
+    end
+
 end
