+++ app/serializers/forecast_serializer.rb
@@ -1,16 +1,6 @@
 class ForecastSerializer
   include JSONAPI::Serializer
-  attributes :id,
-              :type,
-              :sunrise,
-              :sunset,
-              :temperature,
-              :feels_like,
-              :humidity,
-              :uvi,
-              :visibility,
-              :conditions,
-              :icon,
+  attributes :current_weather,
               :hourly_weather,
               :daily_weather
 end
